apps/api/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
celery[redis]==5.4.0
redis==5.2.1
psycopg[binary]==3.2.3
python-dotenv==1.0.1
django-ratelimit==4.1.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1
bandit==1.8.0
